requirements.txt
streamlit
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
cloud-sql-python-connector[pg8000]
streamlit-autorefresh
names
supervisor
google-cloud-logging



Flask
gunicorn
certifi
